+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- ModbusClient::Command TCP transaction ID matching loop having no overall timeout.
+
 ## [1.3.0] - 2026-08-11
 ### Changed
 - Lock timeout handling.